refactor(cli): extract sandbox logs action
